refactor images
