+++ docker-compose.yaml
@@ -1,5 +1,3 @@
-version: "3"
-
 services:
   postgres:
     image: postgres:14
