+++ web/api/app-profile/delete-images/index.ts
@@ -1,0 +1,5 @@
+import { NextRequest } from "next/server";
+
+export const POST = async (req: NextRequest) => {
+  throw new Error("Not implemented yet.");
+};
