+++ web/api/app-profile/upload-image/index.ts
@@ -1,41 +1,33 @@
-import { getSdk as checkUserInAppDocumentSDK } from "@/legacy/api/images/graphql/checkUserInApp.generated";
-import { errorHasuraQuery, errorNotAllowed } from "@/legacy/backend/errors";
-import { getAPIServiceGraphqlClient } from "@/legacy/backend/graphql";
-import { protectInternalEndpoint } from "@/legacy/backend/utils";
+import { getSdk as checkUserInAppDocumentSDK } from "@/api/app-profile/graphql/checkUserInApp.generated";
+import { errorHasuraQuery, errorNotAllowed } from "@/api/helpers/errors";
+import { getAPIServiceGraphqlClient } from "@/api/helpers/graphql";
+import { protectInternalEndpoint } from "@/api/helpers/utils";
+import { validateRequestSchema } from "@/api/helpers/validate-request-schema";
 import { logger } from "@/lib/logger";
 import { S3Client } from "@aws-sdk/client-s3";
 import { createPresignedPost } from "@aws-sdk/s3-presigned-post";
-import { NextApiRequest, NextApiResponse } from "next";
+import { NextRequest, NextResponse } from "next/server";
+import * as yup from "yup";
 
-type RequestQueryParams = {
-  app_id: string;
-  image_type: string;
-  content_type_ending: string;
-};
+const schema = yup.object({
+  app_id: yup.string().strict().required(),
+  image_type: yup.string().strict().required(),
+  content_type_ending: yup.string().required(),
+});
 
-/**
- * Returns a signed url to upload to the predefined path in S3
- * @param req
- * @param res
- */
-
-// TODO: When we migrate this to new API, should convert to GET
-export const handleImageUpload = async (
-  req: NextApiRequest,
-  res: NextApiResponse,
-) => {
+export const POST = async (req: NextRequest) => {
   try {
-    if (!protectInternalEndpoint(req, res)) {
+    if (!protectInternalEndpoint(req)) {
       return;
     }
 
-    if (req.method !== "POST") {
-      return errorNotAllowed(req.method, res, req);
+    if (!req.method || req.method !== "POST") {
+      return errorNotAllowed(req.method, req);
     }
-    const body = req.body;
+
+    const body = await req.json();
     if (body?.action.name !== "upload_image") {
       return errorHasuraQuery({
-        res,
         req,
         detail: "Invalid action.",
         code: "invalid_action",
@@ -45,7 +37,6 @@ export const handleImageUpload = async (
     const userId = body.session_variables["x-hasura-user-id"];
     if (!userId) {
       return errorHasuraQuery({
-        res,
         req,
         detail: "userId must be set.",
         code: "required",
@@ -56,28 +47,29 @@ export const handleImageUpload = async (
 
     if (!teamId) {
       return errorHasuraQuery({
-        res,
         req,
         detail: "teamId must be set.",
         code: "required",
       });
     }
 
-    // TODO: Use yup to validate input
-    const { app_id, image_type, content_type_ending } =
-      req.query as RequestQueryParams;
-    if (!app_id || !image_type || !content_type_ending) {
+    const { isValid, parsedParams } = await validateRequestSchema({
+      value: Object.fromEntries(req.nextUrl.searchParams),
+      schema,
+    });
+
+    if (!isValid || !parsedParams) {
       return errorHasuraQuery({
-        res,
         req,
-        detail: "app_id, image_type, and content_type_ending must be set.",
-        code: "required",
+        detail: "Invalid request params.",
+        code: "invalid_request",
       });
     }
-    // Check that content_type_ending is png or jpeg
+
+    const { app_id, image_type, content_type_ending } = parsedParams;
+
     if (!["png", "jpeg"].includes(content_type_ending)) {
       return errorHasuraQuery({
-        res,
         req,
         detail: "Content Type is invalid",
         code: "invalid_input",
@@ -96,7 +88,6 @@ export const handleImageUpload = async (
     // Admins and Owners allowed to upload images
     if (userTeam.length === 0) {
       return errorHasuraQuery({
-        res,
         req,
         detail: "App not found.",
         code: "not_found",
@@ -105,17 +96,22 @@ export const handleImageUpload = async (
     if (!process.env.ASSETS_S3_REGION) {
       throw new Error("AWS Region must be set.");
     }
+
     const s3Client = new S3Client({
       region: process.env.ASSETS_S3_REGION,
     });
+
     if (!process.env.ASSETS_S3_BUCKET_NAME) {
       throw new Error("AWS Bucket Name must be set.");
     }
+
     const bucketName = process.env.ASSETS_S3_BUCKET_NAME;
+
     // Standardize JPEG to jpg
     const objectKey = `unverified/${app_id}/${image_type}.${
       content_type_ending === "jpeg" ? "jpg" : content_type_ending
     }`;
+
     const contentType = `image/${content_type_ending}`;
     const signedUrl = await createPresignedPost(s3Client, {
       Bucket: bucketName,
@@ -126,18 +122,20 @@ export const handleImageUpload = async (
         ["eq", "$Content-Type", contentType],
       ],
     });
+
     const { url, fields } = signedUrl;
-    res.status(200).json({
-      url,
+
+    return NextResponse.json({
+      url: url,
       stringifiedFields: JSON.stringify(fields),
     });
-  } catch (error: any) {
+  } catch (error) {
     logger.error("Error uploading image.", { error });
+
     return errorHasuraQuery({
-      res,
       req,
       detail: "Unable to upload image",
-      code: "internal_server_error",
+      code: "internal_error",
     });
   }
 };
