+++ web/api/helpers/graphql.ts
@@ -1,6 +1,10 @@
 import "server-only";
 
-import { generateAPIKeyJWT, generateServiceJWT } from "@/api/helpers/jwts";
+import {
+  generateAPIKeyJWT,
+  generateReviewerJWT,
+  generateServiceJWT,
+} from "@/api/helpers/jwts";
 import { GraphQLClient } from "graphql-request";
 
 /**
@@ -16,10 +20,29 @@ export const getAPIServiceGraphqlClient = async () => {
   });
 };
 
+/**
+ * Used for generated requests
+ * Returns an GraphQLClient to interact with GraphQL's API with a API Key token
+ * @returns
+ */
 export const getAPIKeyGraphqlClient = async (params: { team_id: string }) => {
   return new GraphQLClient(process.env.NEXT_PUBLIC_GRAPHQL_API_URL!, {
     headers: {
       authorization: `Bearer ${await generateAPIKeyJWT(params.team_id)}`,
+    },
+  });
+};
+
+/**
+ * Used for generated requests
+ * Returns an GraphQLClient to interact with GraphQL's API with a reviewer token
+ * See Documentation: https://www.notion.so/worldcoin/Reviewer-Role-Specification-5c43c442735842d7ae57e19823a962fb?pvs=4
+ * @returns
+ */
+export const getAPIReviewerGraphqlClient = async () => {
+  return new GraphQLClient(process.env.NEXT_PUBLIC_GRAPHQL_API_URL!, {
+    headers: {
+      authorization: `Bearer ${await generateReviewerJWT()}`,
     },
   });
 };
