+++ web/api/helpers/utils.ts
@@ -4,6 +4,8 @@ import "server-only";
  * Contains shared utilities that are reused for the Next.js API (backend)
  */
 import crypto from "crypto";
+import { NextRequest } from "next/server";
+import { errorForbidden } from "./errors";
 
 const GENERAL_SECRET_KEY = process.env.GENERAL_SECRET_KEY;
 
@@ -12,6 +14,25 @@ if (!GENERAL_SECRET_KEY) {
     "Improperly configured. `GENERAL_SECRET_KEY` env var must be set!",
   );
 }
+
+/**
+ * Ensures endpoint is properly authenticated using internal token. For interactions between Hasura -> Next.js API
+ * @param req
+ * @param res
+ * @returns
+ */
+export const protectInternalEndpoint = (req: NextRequest): boolean => {
+  const requestHeaders = new Headers(req.headers);
+  if (
+    !process.env.INTERNAL_ENDPOINTS_SECRET ||
+    requestHeaders.get("authorization")?.replace("Bearer ", "") !==
+      process.env.INTERNAL_ENDPOINTS_SECRET
+  ) {
+    errorForbidden(req);
+    return false;
+  }
+  return true;
+};
 
 export const verifyHashedSecret = (
   identifier: string,
@@ -23,4 +44,8 @@ export const verifyHashedSecret = (
   const generated_secret = hmac.digest("hex");
 
   return generated_secret === hashed_secret;
+};
+
+export const getFileExtension = (filename: string): string => {
+  return filename.slice(filename.lastIndexOf("."));
 };
