+++ web/app/api/app-profile/get-app-review-images/route.ts
@@ -1,0 +1,1 @@
+export { POST } from "@/api/app-profile/get-app-review-images";
