+++ web/app/api/app-profile/upload-image/route.ts
@@ -1,0 +1,1 @@
+export { POST } from "@/api/app-profile/upload-image";
