+++ web/app/api/app-profile/verify-app/route.ts
@@ -1,0 +1,1 @@
+export { POST } from "@/api/app-profile/verify-app";
