+++ web/scenes/Portal/Teams/TeamId/Apps/AppId/Profile/hook/use-image.tsx
@@ -100,6 +100,7 @@ export const useImage = () => {
     }
 
     const { url, stringifiedFields } = response.data.upload_image;
+
     const fields = JSON.parse(stringifiedFields);
 
     const formData = new FormData();
